+++ modules/runners/variables.tf
@@ -651,7 +651,7 @@ variable "placement" {
     group_id                = optional(string)
     group_name              = optional(string)
     host_id                 = optional(string)
-    host_resource_group_arn = optional(number)
+    host_resource_group_arn = optional(string)
     spread_domain           = optional(string)
     tenancy                 = optional(string)
     partition_number        = optional(number)
@@ -699,6 +699,22 @@ variable "enable_on_demand_failover_for_errors" {
   description = "Enable on-demand failover. For example to fall back to on demand when no spot capacity is available the variable can be set to `InsufficientInstanceCapacity`. When not defined the default behavior is to retry later."
   type        = list(string)
   default     = []
+}
+
+variable "scale_errors" {
+  description = "List of aws error codes that should trigger retry during scale up. This list will replace the default errors defined in the variable `defaultScaleErrors` in https://github.com/github-aws-runners/terraform-aws-github-runner/blob/main/lambdas/functions/control-plane/src/aws/runners.ts"
+  type        = list(string)
+  default = [
+    "UnfulfillableCapacity",
+    "MaxSpotInstanceCountExceeded",
+    "TargetCapacityLimitExceededException",
+    "RequestLimitExceeded",
+    "ResourceLimitExceeded",
+    "MaxSpotInstanceCountExceeded",
+    "MaxSpotFleetRequestCountExceeded",
+    "InsufficientInstanceCapacity",
+    "InsufficientCapacityOnHost",
+  ]
 }
 
 variable "lambda_tags" {